modularize for setup.py
