+++ pyreil/x86/logic.py
@@ -14,7 +14,7 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-"""reil.x86.logic - x86 and x86_64 translators
+"""pyreil.x86.logic - x86 and x86_64 translators
 
 This module generates REIL (reverse engineering intermediate language)
 IL from x86 and x86_64 machine code.
@@ -23,13 +23,12 @@ This file is responsible for translation of basic logical instructions
 such as and, or, xor
 """
 
-import reil.error
-from reil.shorthand import *
+from pyreil.x86 import operand
+from pyreil.x86 import conditional
+from pyreil.x86.utilities import *
 
-import reil.x86.conditional as conditional
-import reil.x86.operand as operand
-from reil.x86.utilities import *
-
+import pyreil.error
+from pyreil.shorthand import *
 
 # Helpers
