+++ pyreil/x86/memory.py
@@ -14,7 +14,7 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-"""reil.x86.memory - x86 and x86_64 translators
+"""pyreil.x86.memory - x86 and x86_64 translators
 
 This module generates REIL (reverse engineering intermediate language)
 IL from x86 and x86_64 machine code.
@@ -25,15 +25,12 @@ such as lea, mov, push, pop
 
 import capstone
 
-import reil.error
-from reil import *
-from reil.shorthand import *
+from pyreil.shorthand import *
 
-import reil.x86.arithmetic as arithmetic
-import reil.x86.conditional as conditional
-import reil.x86.operand as operand
-from reil.x86.utilities import *
-
+from pyreil.x86 import operand
+from pyreil.x86 import arithmetic
+from pyreil.x86 import conditional
+from pyreil.x86.utilities import *
 
 def conditional_mov(ctx, i, condition):
     c = conditional.condition(ctx, condition)
